# internal/handlers/grid.go
package handlers

import (
	"log"
	"net/http"

	"github.com/daniele/web-app-caa/internal/config"
	"github.com/daniele/web-app-caa/internal/models"
	"github.com/daniele/web-app-caa/internal/services"
	"github.com/daniele/web-app-caa/internal/utils/token"

	"github.com/gin-gonic/gin"
)

// GridHandlers handles grid-related requests
type GridHandlers struct {
	gridService *services.GridService
	userService *services.UserService
	cfg         *config.Config
}

// NewGridHandlers creates a new GridHandlers instance
func NewGridHandlers(cfg *config.Config) *GridHandlers {
	return &GridHandlers{
		gridService: services.NewGridService(),
		userService: services.NewUserService(),
		cfg:         cfg,
	}
}

// Setup handles grid setup request
// @Summary Setup grid
// @Description Initialize a grid with the specified type (simplified, empty, or default)
// @Tags Grid
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param request body models.SetupRequest true "Setup request"
// @Success 200 {object} models.SuccessResponse
// @Failure 400 {object} models.ErrorResponse
// @Failure 401 {object} models.ErrorResponse
// @Failure 500 {object} models.ErrorResponse
// @Router /setup [post]
func (h *GridHandlers) Setup(c *gin.Context) {
	// Extract user ID from token
	userID, err := token.ExtractUserIDFromContext(c)
	if err != nil {
		log.Printf("[SETUP] Error extracting user ID from token: %v", err)
		c.JSON(http.StatusUnauthorized, gin.H{"error": "Invalid token"})
		return
	}

	log.Printf("[SETUP] Setup request started for userId: %s", userID)

	var req models.SetupRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"message": "GridType is required."})
		return
	}

	log.Printf("[SETUP] Grid type requested: %s for userId: %d", req.GridType, userID)

	// Determine the grid structure
	var selectedGrid map[string][]models.GridItemResponse
	switch req.GridType {
	case "simplified":
		selectedGrid = services.GetSimplifiedGrid(h.cfg)
		log.Printf("[SETUP] Selected simplified grid")
	case "empty":
		defaultGrid := services.GetDefaultGrid(h.cfg)
		selectedGrid = map[string][]models.GridItemResponse{
			"home":           {},
			"systemControls": defaultGrid["systemControls"],
		}
		log.Printf("[SETUP] Selected empty grid with system controls")
	case "default":
	default:
		selectedGrid = services.GetDefaultGrid(h.cfg)
		log.Printf("[SETUP] Selected default grid")
	}

	// Save grid
	log.Printf("[SETUP] Saving grid to database for userId: %d", userID)
	if err := h.gridService.SaveGrid(selectedGrid, userID); err != nil {
		log.Printf("[SETUP] Error saving grid: %v", err)
		c.JSON(http.StatusInternalServerError, gin.H{
			"message": "Error saving setup.",
			"error":   err.Error(),
		})
		return
	}

	// Update user status
	log.Printf("[SETUP] Updating user status to 'active' for userId: %d", userID)
	if err := h.userService.UpdateUserStatus(userID, "active"); err != nil {
		log.Printf("[SETUP] Error updating user status: %v", err)
		c.JSON(http.StatusInternalServerError, gin.H{
			"message": "Error updating user status.",
			"error":   err.Error(),
		})
		return
	}

	log.Printf("[SETUP] Setup completed successfully for userId: %s", userID)
	c.JSON(http.StatusOK, gin.H{"message": "Setup complete. Grid saved."})
}

// CompleteSetup handles setup completion request
// @Summary Complete setup
// @Description Mark the user setup as complete
// @Tags Grid
// @Produce json
// @Security BearerAuth
// @Success 200 {object} models.SuccessResponse
// @Failure 401 {object} models.ErrorResponse
// @Failure 500 {object} models.ErrorResponse
// @Router /complete-setup [post]
func (h *GridHandlers) CompleteSetup(c *gin.Context) {
	// Extract user ID from token
	userID, err := token.ExtractUserIDFromContext(c)
	if err != nil {
		log.Printf("[COMPLETE-SETUP] Error extracting user ID from token: %v", err)
		c.JSON(http.StatusUnauthorized, gin.H{"error": "Invalid token"})
		return
	}

	log.Printf("[COMPLETE-SETUP] Complete setup request started for userId: %s", userID)

	// Update user status to completed
	if err := h.userService.UpdateUserStatus(userID, "completed"); err != nil {
		log.Printf("[COMPLETE-SETUP] Error updating user status: %v", err)
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to complete setup"})
		return
	}

	log.Printf("[COMPLETE-SETUP] Setup completed successfully for userId: %s", userID)
	c.JSON(http.StatusOK, gin.H{"message": "Setup completed successfully"})
}

// GetGrid retrieves the entire grid for a user
// @Summary Get grid
// @Description Retrieve the complete grid configuration for the current user
// @Tags Grid
// @Produce json
// @Security BearerAuth
// @Success 200 {object} models.GridResponse
// @Failure 401 {object} models.ErrorResponse
// @Failure 500 {object} models.ErrorResponse
// @Router /grid [get]
func (h *GridHandlers) GetGrid(c *gin.Context) {
	userID, err := token.ExtractUserIDFromContext(c)
	if err != nil {
		log.Printf("[ERROR] Error extracting user ID from token: %v", err)
		c.JSON(http.StatusUnauthorized, gin.H{"error": "Invalid token"})
		return
	}

	log.Printf("[GET-GRID] Retrieving grid for userId: %s", userID)

	gridData, err := h.gridService.GetGrid(userID)
	if err != nil {
		log.Printf("[GET-GRID] Error reading from database: %v", err)
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Error reading from database."})
		return
	}

	if gridData == nil {
		gridData = make(map[string][]models.GridItemResponse)
	}

	log.Printf("[GET-GRID] Grid retrieved successfully for userId: %d, has data: %t",
		userID, len(gridData) > 0)

	c.JSON(http.StatusOK, gridData)
}

// SaveGrid saves a full grid (updates from client)
// @Summary Save grid
// @Description Save the complete grid configuration
// @Tags Grid
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param grid body models.GridResponse true "Grid data to save"
// @Success 200 {object} models.SuccessResponse
// @Failure 400 {object} models.ErrorResponse
// @Failure 401 {object} models.ErrorResponse
// @Failure 500 {object} models.ErrorResponse
// @Router /grid [post]
func (h *GridHandlers) SaveGrid(c *gin.Context) {
	userID, err := token.ExtractUserIDFromContext(c)
	if err != nil {
		log.Printf("[ERROR] Error extracting user ID from token: %v", err)
		c.JSON(http.StatusUnauthorized, gin.H{"error": "Invalid token"})
		return
	}

	log.Printf("[SAVE-GRID] Saving grid for userId: %s", userID)

	var gridData map[string][]models.GridItemResponse
	if err := c.ShouldBindJSON(&gridData); err != nil {
		log.Printf("[SAVE-GRID] Invalid grid data: %v", err)
		c.JSON(http.StatusBadRequest, gin.H{"message": "Invalid grid data."})
		return
	}

	if err := h.gridService.SaveGrid(gridData, userID); err != nil {
		log.Printf("[SAVE-GRID] Error writing to database: %v", err)
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Error saving to database."})
		return
	}

	log.Printf("[SAVE-GRID] Grid saved successfully for userId: %d", userID)
	c.JSON(http.StatusOK, gin.H{"message": "Grid saved successfully!"})
}

// AddItem adds a new item to the grid
// @Summary Add grid item
// @Description Add a new item to a specific category in the grid
// @Tags Grid
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param request body models.AddItemRequest true "Add item request"
// @Success 201 {object} models.GridItemResponse
// @Failure 400 {object} models.ErrorResponse
// @Failure 401 {object} models.ErrorResponse
// @Failure 500 {object} models.ErrorResponse
// @Router /grid/item [post]
func (h *GridHandlers) AddItem(c *gin.Context) {
	userID, err := token.ExtractUserIDFromContext(c)
	if err != nil {
		log.Printf("[ERROR] Error extracting user ID from token: %v", err)
		c.JSON(http.StatusUnauthorized, gin.H{"error": "Invalid token"})
		return
	}

	log.Printf("[ADD-ITEM] Adding new item for userId: %s", userID)

	var req models.AddItemRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		log.Printf("[ADD-ITEM] Invalid request payload: %v", err)
		c.JSON(http.StatusBadRequest, gin.H{
			"message": "Item data and parent category are required.",
		})
		return
	}

	log.Printf("[ADD-ITEM] Item data: ID=%s, ParentCategory=%s",
		req.Item.ID, req.ParentCategory)

	newItem, err := h.gridService.AddItem(req.Item, req.ParentCategory, userID)
	if err != nil {
		log.Printf("[ADD-ITEM] Error adding item to database: %v", err)
		c.JSON(http.StatusInternalServerError, gin.H{
			"message": "Error adding item.",
			"error":   err.Error(),
		})
		return
	}

	log.Printf("[ADD-ITEM] Item added successfully for userId: %d: %s", userID, newItem.ID)
	c.JSON(http.StatusCreated, newItem)
}

// UpdateItem updates an existing item
// @Summary Update grid item
// @Description Update an existing item in the grid
// @Tags Grid
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path string true "Item ID"
// @Param item body models.GridItemResponse true "Updated item data"
// @Success 200 {object} models.SuccessResponse
// @Failure 400 {object} models.ErrorResponse
// @Failure 401 {object} models.ErrorResponse
// @Failure 404 {object} models.ErrorResponse
// @Failure 500 {object} models.ErrorResponse
// @Router /grid/item/{id} [put]
func (h *GridHandlers) UpdateItem(c *gin.Context) {
	userID, err := token.ExtractUserIDFromContext(c)
	if err != nil {
		log.Printf("[ERROR] Error extracting user ID from token: %v", err)
		c.JSON(http.StatusUnauthorized, gin.H{"error": "Invalid token"})
		return
	}

	itemID := c.Param("id")

	var updateData models.GridItemResponse
	if err := c.ShouldBindJSON(&updateData); err != nil {
		log.Printf("[UPDATE-ITEM] Invalid update data: %v", err)
		c.JSON(http.StatusBadRequest, gin.H{"message": "Invalid update data."})
		return
	}

	log.Printf("[UPDATE-ITEM] Update data for item %s", itemID)

	if err := h.gridService.UpdateItem(itemID, updateData, userID); err != nil {
		log.Printf("[UPDATE-ITEM] Error updating item in database: %v", err)
		c.JSON(http.StatusInternalServerError, gin.H{
			"message": "Error updating item.",
			"error":   err.Error(),
		})
		return
	}

	log.Printf("[UPDATE-ITEM] Item updated successfully: %s for userId: %d", itemID, userID)
	c.JSON(http.StatusOK, gin.H{"message": "Item updated successfully!"})
}

// DeleteItem deletes an item
// @Summary Delete grid item
// @Description Delete an existing item from the grid
// @Tags Grid
// @Produce json
// @Security BearerAuth
// @Param id path string true "Item ID"
// @Success 200 {object} models.SuccessResponse
// @Failure 401 {object} models.ErrorResponse
// @Failure 404 {object} models.ErrorResponse
// @Failure 500 {object} models.ErrorResponse
// @Router /grid/item/{id} [delete]
func (h *GridHandlers) DeleteItem(c *gin.Context) {
	userID, err := token.ExtractUserIDFromContext(c)
	if err != nil {
		log.Printf("[ERROR] Error extracting user ID from token: %v", err)
		c.JSON(http.StatusUnauthorized, gin.H{"error": "Invalid token"})
		return
	}

	itemID := c.Param("id")

	log.Printf("[DELETE-ITEM] Deleting item %s for userId: %s", itemID, userID)

	var req struct {
		CategoryTarget string `json:"categoryTarget"`
	}
	if err := c.ShouldBindJSON(&req); err != nil {
		log.Printf("[DELETE-ITEM] Warning: failed to parse optional categoryTarget: %v", err)
	}

	if req.CategoryTarget != "" {
		log.Printf("[DELETE-ITEM] Category target specified: %s", req.CategoryTarget)
	}

	// Delete the item
	if err := h.gridService.DeleteItem(itemID, userID); err != nil {
		log.Printf("[DELETE-ITEM] Error deleting item from database: %v", err)
		c.JSON(http.StatusInternalServerError, gin.H{
			"message": "Error deleting item.",
			"error":   err.Error(),
		})
		return
	}

	log.Printf("[DELETE-ITEM] Item deleted: %s for userId: %s", itemID, userID)

	// Delete category contents if specified
	if req.CategoryTarget != "" {
		log.Printf("[DELETE-ITEM] Deleting category contents for: %s", req.CategoryTarget)
		if err := h.gridService.DeleteCategoryContents(req.CategoryTarget, userID); err != nil {
			log.Printf("[DELETE-ITEM] Error deleting category contents: %v", err)
			// Don't return error here as the main item was already deleted
		} else {
			log.Printf("[DELETE-ITEM] Category contents deleted for: %s", req.CategoryTarget)
		}
	}

	c.JSON(http.StatusOK, gin.H{"message": "Item deleted successfully!"})
}